Moved useEffect logic to useCheckout hook and minor changes
